LLM/requirements.txt
groq
sentence-transformers
pandas
python-dotenv
tqdm
scikit-learn 
numpy
fastapi
uvicorn
boto3
google-generativeai
pydantic
google-cloud-aiplatform